mixin config plugin shenanigans
