+++ neoforge/src/main/java/org/geysermc/floodgate/shared/util/neoforge/ModMixinConfigPluginImpl.java
@@ -1,0 +1,13 @@
+package org.geysermc.floodgate.shared.util.neoforge;
+
+import net.neoforged.fml.loading.LoadingModList;
+
+public class ModMixinConfigPluginImpl {
+    private static boolean isGeyserLoaded() {
+        return LoadingModList.get().getModFileById("geyser_neoforge") != null;
+    }
+
+    private static boolean applyProxyFix() {
+        return false;
+    }
+}
