+++ shared/src/main/java/org/geysermc/floodgate/shared/util/ModMixinConfigPlugin.java
@@ -1,5 +1,6 @@
 package org.geysermc.floodgate.shared.util;
 
+import dev.architectury.injectables.annotations.ExpectPlatform;
 import org.objectweb.asm.tree.ClassNode;
 import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
 import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
@@ -7,7 +8,7 @@ import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
 import java.util.List;
 import java.util.Set;
 
-public abstract class ModMixinConfigPlugin implements IMixinConfigPlugin {
+public class ModMixinConfigPlugin implements IMixinConfigPlugin {
 
     @Override
     public void onLoad(String mixinPackage) {
@@ -19,7 +20,15 @@ public abstract class ModMixinConfigPlugin implements IMixinConfigPlugin {
     }
 
     @Override
-    public abstract boolean shouldApplyMixin(String targetClassName, String mixinClassName);
+    public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
+        if (mixinClassName.equals("org.geysermc.floodgate.mixin.ClientIntentionPacketMixin")) {
+            return applyProxyFix();
+        }
+        if (mixinClassName.equals("org.geysermc.floodgate.mixin.GeyserModInjectorMixin")) {
+            return isGeyserLoaded();
+        }
+        return true;
+    }
 
     @Override
     public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {
@@ -36,5 +45,15 @@ public abstract class ModMixinConfigPlugin implements IMixinConfigPlugin {
 
     @Override
     public void postApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
+    }
+
+    @ExpectPlatform
+    private static boolean isGeyserLoaded() {
+        throw new IllegalStateException("isGeyserLoaded is not implemented!");
+    }
+
+    @ExpectPlatform
+    private static boolean applyProxyFix() {
+        throw new IllegalStateException("applyProxyFix is not implemented!");
     }
 }
